+++ packages/cli/src/pages/api/__integration__/util/apiKeys.ts
@@ -1,4 +1,4 @@
-import { cliUrl, fetch } from "./index";
+import { Api, cliUrl } from "./index";
 
 export async function apiCreateApiKey() {
   const instance = new ApiPostApiKeys();
@@ -9,9 +9,7 @@ export async function apiGetApiKeys() {
   return instance.perform();
 }
 
-export class ApiGetApiKeys {
-  response?: Awaited<ReturnType<typeof fetch>>;
-  fetchData: any;
+export class ApiGetApiKeys extends Api {
   path = cliUrl("/api/apiKeys");
 
   defaultFetchData = {
@@ -20,31 +18,8 @@ export class ApiGetApiKeys {
       "Content-Type": "application/json",
     },
   };
-
-  async perform() {
-    this.fetchData = {
-      ...this.defaultFetchData,
-    };
-
-    this.response = await fetch(this.path, this.fetchData);
-
-    return this;
-  }
 }
 
-export class ApiPostApiKeys {
-  response?: Awaited<ReturnType<typeof fetch>>;
-  fetchData = {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-  };
+export class ApiPostApiKeys extends Api {
   path = cliUrl("/api/apiKeys");
-
-  async perform() {
-    this.response = await fetch(this.path, this.fetchData);
-
-    return this;
-  }
 }
