+++ packages/cli/src/pages/api/__integration__/util/login.ts
@@ -1,4 +1,4 @@
-import { cliUrl, fetch } from ".";
+import { Api, cliUrl } from ".";
 import { apiCreateUser } from "./createUser";
 
 interface LoginFormData {
@@ -22,37 +22,6 @@ export async function apiLogin() {
   expect(apiLoginResponse.status).toBe(201);
 }
 
-export class ApiLogin {
-  response?: Awaited<ReturnType<typeof fetch>>;
-  fetchData: any;
-  formData: Partial<LoginFormData>;
+export class ApiLogin extends Api<LoginFormData> {
   path = cliUrl("/api/session");
-
-  defaultFormData = {};
-
-  defaultFetchData = {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-  };
-
-  constructor() {
-    this.formData = this.defaultFormData;
-  }
-
-  async perform() {
-    this.fetchData = {
-      ...this.defaultFetchData,
-      body: JSON.stringify(this.formData),
-    };
-
-    this.response = await fetch(this.path, this.fetchData);
-
-    return this;
-  }
-
-  updateFormData(data: Partial<LoginFormData>) {
-    this.formData = { ...this.formData, ...data };
-  }
 }
