+++ packages/cli/src/pages/api/__integration__/util/logout.ts
@@ -1,23 +1,10 @@
-import { cliUrl, fetch } from "./index";
+import { Api, cliUrl } from "./index";
 
 export function apiLogout() {
   const instance = new ApiLogout();
   return instance.perform();
 }
 
-export class ApiLogout {
-  response?: Awaited<ReturnType<typeof fetch>>;
-  fetchData = {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-  };
+export class ApiLogout extends Api {
   path = cliUrl("/api/logout");
-
-  async perform() {
-    this.response = await fetch(this.path, this.fetchData);
-
-    return this;
-  }
 }
